Extract integration test configuration into a shared Gradle convention plugin and apply it to `core` and `client` modules.
