+++ core/build.gradle.kts
@@ -1,7 +1,4 @@
 import java.net.URI
-import org.gradle.api.tasks.testing.TestDescriptor
-import org.gradle.api.tasks.testing.TestListener
-import org.gradle.api.tasks.testing.TestResult
 import org.gradle.api.tasks.testing.logging.TestLogEvent
 import org.gradle.external.javadoc.StandardJavadocDocletOptions
 
@@ -12,6 +9,7 @@ plugins {
     alias(libs.plugins.shadow)
     alias(libs.plugins.vt.publish)
     jacoco
+    id("integration-test-conventions")
 }
 
 repositories {
@@ -142,16 +140,6 @@ java {
     targetCompatibility = JavaVersion.toVersion(jvmVersion)
 }
 
-val integrationTestSourceSet = sourceSets.create("integrationTest") {
-    compileClasspath += sourceSets["main"].output
-    runtimeClasspath += sourceSets["main"].output
-}
-
-configurations["integrationTestImplementation"]
-    .extendsFrom(configurations["testImplementation"])
-configurations["integrationTestRuntimeOnly"]
-    .extendsFrom(configurations["testRuntimeOnly"])
-
 val integrationTestResources : File = File(project.layout.buildDirectory.get().asFile, "integrationTestResources")
 
 val downloadIntegrationTestResources = tasks.register("downloadIntegrationTestResources") {
@@ -187,46 +175,10 @@ val downloadIntegrationTestResources = tasks.register("downloadIntegrationTestRe
     }
 }
 
-// Add a task to run the integration tests
-val integrationTest = tasks.register<Test>("integrationTest") {
+tasks.named<Test>("integrationTest") {
     dependsOn(downloadIntegrationTestResources)
-    group = "Verification"
     description = "Runs integration tests in core/src/integrationTest"
-    shouldRunAfter(tasks.test)
-    testClassesDirs = integrationTestSourceSet.output.classesDirs
-    classpath = integrationTestSourceSet.runtimeClasspath
     environment("integrationTestingDir", File(integrationTestResources.absolutePath, "testset"))
-
-    useJUnitPlatform()
-
-    testLogging {
-        events(TestLogEvent.PASSED, TestLogEvent.SKIPPED, TestLogEvent.FAILED)
-        exceptionFormat = org.gradle.api.tasks.testing.logging.TestExceptionFormat.FULL
-        showStandardStreams = true
-    }
-
-    val executedClasses = mutableSetOf<String>()
-    addTestListener(
-        object : TestListener {
-            override fun beforeSuite(suite: TestDescriptor) {}
-            override fun beforeTest(testDescriptor: TestDescriptor) {}
-            override fun afterTest(testDescriptor: TestDescriptor, result: TestResult) {
-                testDescriptor.className?.let { executedClasses.add(it) }
-            }
-
-            override fun afterSuite(suite: TestDescriptor, result: TestResult) {
-                if (suite.parent == null) {
-                    println(
-                        "Integration tests: ${result.testCount} invocations " +
-                            "from ${executedClasses.size} classes; " +
-                            "${result.successfulTestCount} passed, " +
-                            "${result.failedTestCount} failed, " +
-                            "${result.skippedTestCount} skipped"
-                    )
-                }
-            }
-        },
-    )
 }
 
 tasks.jacocoTestReport {
